organise firebase services for use
